backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
fastapi==0.115.2
h11==0.14.0
idna==3.10
numpy==2.1.2
pillow==11.0.0
pydantic==2.9.2
pydantic_core==2.23.4
python-multipart==0.0.12
sniffio==1.3.1
starlette==0.40.0
typing_extensions==4.12.2
uvicorn==0.32.0
